+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Return `[pid, exit_status]` from `Gyro::Child#await`
+
 0.25 2020-01-10
 ---------------